+++ pretext/cli.py
@@ -27,6 +27,7 @@ from . import (
     core,
     constants,
     plastex,
+    server,
     VERSION,
     CORE_COMMIT,
 )
@@ -783,18 +784,23 @@ def view(
     """
 
     # pretext view -s should immediately stop the server and do nothing else.
+    if utils.cannot_find_project(task="view the output for"):
+        return
+    project = Project.parse()
+
     if stop_server:
         try:
+            projectHash = utils.hash_path(project.abspath())
+            current_server = server.active_server_for_path_hash(projectHash)
             log.info("\nStopping server.")
-            utils.stop_server()
+            if current_server:
+                current_server.terminate()
         except Exception as e:
             log.warning("Failed to stop server.")
             log.debug(e, exc_info=True)
         finally:
             return
-    if utils.cannot_find_project(task="view the output for"):
-        return
-    project = Project.parse()
+
     try:
         target = project.get_target(name=target_name, log_info_for_none=not stage)
     except AssertionError as e:
@@ -848,8 +854,26 @@ def view(
             webbrowser.open(url)
         return
     # Start server if there isn't one running already:
-    used_port = utils.active_server_port()
-    if restart_server or (port != used_port) or (used_port is None):
+    projectHash = utils.hash_path(project.abspath())
+    current_server = server.active_server_for_path_hash(projectHash)
+    if restart_server and current_server is not None:
+        log.info(
+            f"Terminating existing server {current_server.pid} on port {current_server.port}"
+        )
+        current_server.terminate()
+        current_server = None
+    # Double check that the current server really is active:
+    if current_server is not None and current_server.isActiveServer():
+        url_base = utils.url_for_access(access=access, port=current_server.port)
+        url = url_base + url_path
+        log.info(f"Server is already available at {url_base}")
+        if no_launch:
+            log.info(f"The {target_name} is available at {url}")
+        else:
+            log.info(f"Now opening browser for {target_name} at {url}")
+            webbrowser.open(url)
+    # otherwise, start a server
+    else:
         log.info(
             f"Now preparing local server to preview your project directory `{project.abspath()}`."
         )
@@ -861,50 +885,21 @@ def view(
         )
         log.info("  personal computer.)")
         log.info("")
-        # First terminate any existing server using this port
-        if used_port == port:
-            try:
-                utils.stop_server(used_port)
-            except Exception as e:
-                log.warning("Failed to stop server.")
-                log.debug(e, exc_info=True)
-                pass
-        # Start the new server
-        server = project.server_process(
-            access=access,
-            port=port,
-        )
-        server.start()
 
-        # Now get the updated port in case we had to pick a new one
-        actual_port = utils.active_server_port() or port
-        # set the url
-        url_base = utils.url_for_access(access=access, port=actual_port)
-        url = url_base + url_path
-        log.info(f"Server will soon be available at {url_base}")
-        if no_launch:
-            log.info(f"The {target_name} will be available at {url}")
-        else:
-            # SECONDS = 2
-            log.info(f"Opening browser for {target_name} at {url}")
-            # time.sleep(SECONDS)
-            webbrowser.open(url)
-        try:
-            while server.is_alive():
-                time.sleep(1)
-        except KeyboardInterrupt:
-            log.info("Stopping server.")
-            server.terminate()
-            return
-    else:
-        url_base = utils.url_for_access(access=access, port=used_port)
-        url = url_base + url_path
-        log.info(f"Server is already available at {url_base}")
-        if no_launch:
-            log.info(f"The {target_name} is available at {url}")
-        else:
-            log.info(f"Now opening browser for {target_name} at {url}")
-            webbrowser.open(url)
+        def callback(actual_port: int) -> None:
+            url_base = utils.url_for_access(access=access, port=actual_port)
+            url = url_base + url_path
+            log.info(f"Server will soon be available at {url_base}")
+            if no_launch:
+                log.info(f"The {target_name} will be available at {url}")
+            else:
+                # SECONDS = 2
+                log.info(f"Opening browser for {target_name} at {url}")
+                # time.sleep(SECONDS)
+                webbrowser.open(url)
+
+        log.info("starting server ...")
+        server.start_server(project.abspath(), access, port, callback)
 
 
 # pretext deploy