+++ CHANGELOG.md
@@ -9,6 +9,9 @@ Instructions: Add a subsection under `[Unreleased]` for additions, fixes, change
 
 ## [Unreleased]
 
+### Changed
+
+- `pretext view` now reuses the current local server correctly, and starts different local servers for different projects.  This allows you to correctly view multiple projects at the same time.
 
 ## [2.10.1] - 2024-12-10